+++ berkeley-function-call-leaderboard/bfcl_eval/constants/supported_models.py
@@ -25,8 +25,6 @@ SUPPORTED_MODELS = [
     "gpt-4.1-mini-2025-04-14",
     "gpt-4.1-nano-2025-04-14-FC",
     "gpt-4.1-nano-2025-04-14",
-    "o1-2024-12-17-FC",
-    "o1-2024-12-17",
     "o3-2025-04-16",
     "o3-2025-04-16-FC",
     "o4-mini-2025-04-16",
