+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/api_inference/deepseek.py
@@ -3,7 +3,7 @@ import os
 import time
 from typing import Any
 
-from bfcl_eval.model_handler.api_inference.openai import OpenAICompletionsHandler
+from bfcl_eval.model_handler.api_inference.openai_completion import OpenAICompletionsHandler
 from bfcl_eval.model_handler.model_style import ModelStyle
 from bfcl_eval.model_handler.utils import (
     combine_consecutive_user_prompts,
