+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/api_inference/functionary.py
@@ -1,4 +1,4 @@
-from bfcl_eval.model_handler.api_inference.openai import OpenAICompletionsHandler
+from bfcl_eval.model_handler.api_inference.openai_completion import OpenAICompletionsHandler
 from bfcl_eval.model_handler.model_style import ModelStyle
 from openai import OpenAI
