+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/api_inference/gogoagent.py
@@ -1,6 +1,6 @@
 import os
 
-from bfcl_eval.model_handler.api_inference.openai import OpenAICompletionsHandler
+from bfcl_eval.model_handler.api_inference.openai_completion import OpenAICompletionsHandler
 from openai import OpenAI
