+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/api_inference/grok.py
@@ -1,6 +1,6 @@
 import os
 
-from bfcl_eval.model_handler.api_inference.openai import OpenAICompletionsHandler
+from bfcl_eval.model_handler.api_inference.openai_completion import OpenAICompletionsHandler
 from openai import OpenAI
 from overrides import override
 from typing import Any
