+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/api_inference/nvidia.py
@@ -1,7 +1,7 @@
 import os
 
 from bfcl_eval.model_handler.model_style import ModelStyle
-from bfcl_eval.model_handler.api_inference.openai import OpenAICompletionsHandler
+from bfcl_eval.model_handler.api_inference.openai_completion import OpenAICompletionsHandler
 from bfcl_eval.model_handler.utils import (
     ast_parse,
     combine_consecutive_user_prompts,
