+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/api_inference/qwen.py
@@ -1,7 +1,7 @@
 import os
 from typing import Any
 
-from bfcl_eval.model_handler.api_inference.openai import OpenAICompletionsHandler
+from bfcl_eval.model_handler.api_inference.openai_completion import OpenAICompletionsHandler
 from bfcl_eval.model_handler.model_style import ModelStyle
 from openai import OpenAI
 from overrides import override
